+++ cpp/tests/main.cpp
@@ -20,45 +20,8 @@
 
 #include <gtest/gtest.h>
 
-//#import "math.h"
-
-namespace {
-
-/**
- * The fixture for testing class Foo.
- */
-class MathTest : public ::testing::Test {
- protected:
-  // You can remove any or all of the following functions if their bodies would
-  // be empty.
-
-  MathTest() {
-     // You can do set-up work for each test here.
-  }
-
-  ~MathTest() override {
-     // You can do clean-up work that doesn't throw exceptions here.
-  }
-
-  // If the constructor and destructor are not enough for setting up
-  // and cleaning up each test, you can define the following methods:
-
-  void SetUp() override {
-     // Code here will be called immediately after the constructor (right
-     // before each test).
-  }
-
-  void TearDown() override {
-     // Code here will be called immediately after each test (right
-     // before the destructor).
-  }
-
-  // Class members declared here can be used by all tests in the test suite
-  // for Foo.
-};
-}
-
-// Tests that the Foo::Bar() method does Abc.
-TEST_F(MathTest, Vector3) {
-  EXPECT_EQ(42, 42);
+int main(int argc, char **argv) {
+  // Runs all the tests specified in *-test.cpp files that have been registered using TEST or TEST_F macros
+  ::testing::InitGoogleTest(&argc, argv);
+  return RUN_ALL_TESTS();
 }
