+++ cpp/tests/math.gl/core-test.cpp
@@ -1,0 +1,72 @@
+// Copyright (c) 2020, Unfolded Inc
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+#include "math.gl/core.h"
+
+#include <gtest/gtest.h>
+
+namespace {
+
+/**
+ * The fixture for testing class Foo.
+ */
+class MathTest : public ::testing::Test {
+ protected:
+  // You can remove any or all of the following functions if their bodies would
+  // be empty.
+
+  MathTest() {
+    // You can do set-up work for each test here.
+  }
+
+  ~MathTest() override {
+    // You can do clean-up work that doesn't throw exceptions here.
+  }
+
+  // If the constructor and destructor are not enough for setting up
+  // and cleaning up each test, you can define the following methods:
+
+  void SetUp() override {
+    // Code here will be called immediately after the constructor (right
+    // before each test).
+  }
+
+  void TearDown() override {
+    // Code here will be called immediately after each test (right
+    // before the destructor).
+  }
+
+  // Class members declared here can be used by all tests in the test suite
+  // for Foo.
+};
+
+// Tests Vector3 + operand
+TEST_F(MathTest, Vector3Sum) {
+  mathgl::Vector3 vector1{0.1, 0.3, 0.5};
+  mathgl::Vector3 vector2{0.6, 0.2, 0.1};
+
+  mathgl::Vector3 vectorSum = vector1 + vector2;
+
+  EXPECT_FLOAT_EQ(vectorSum.x, 0.7);
+  EXPECT_FLOAT_EQ(vectorSum.y, 0.5);
+  EXPECT_FLOAT_EQ(vectorSum.z, 0.6);
+}
+
+}  // namespace
